backend/requirements.txt
fastapi
uvicorn
python-dotenv
motor
requests
openai
groq
google-generativeai
